implement emailVerificationService
