+++ server/.env.example
@@ -4,6 +4,7 @@ DATABASE_URL="postgresql://postgres:postgres@localhost:5432/task_tracker"
 ROOT_URL="http://localhost:5000"
 ACCESS_TOKEN_SECRET="random_string_of_characters_numbers_and_symbols"
 VERIFICATION_TOKEN_SECRET="another_random_string_of_characters_numbers_and_symbols"
+EMAIL_SERVICE="Gmail"
 EMAIL_ADDRESS="noreply@service.com"
 EMAIL_PASSWORD="password"
 EMAIL_NAME="No Reply"
